fix download tests

+++ localtron/services/download/types/downloadtypes.go
@@ -32,8 +32,8 @@ type DownloadDetails struct {
 	FileName        string   `json:"fileName"`
 	Dir             *string  `json:"dir,omitempty"`
 	Progress        *float64 `json:"progress,omitempty"`
-	DownloadedBytes int      `json:"downloadedBytes"`
-	FullFileSize    *int     `json:"fullFileSize"`
+	DownloadedBytes int64    `json:"downloadedBytes"`
+	FullFileSize    *int64   `json:"fullFileSize"`
 	Status          string   `json:"status"`
 	FilePath        *string  `json:"filePath,omitempty"`
 	Paused          *bool    `json:"paused,omitempty"`
